benchdnn: consolidate execution_graph in a single file
