+++ tests/benchdnn/benchdnn.cpp
@@ -78,8 +78,6 @@ int test_start {0};
 bool attr_same_pd_check {false};
 bool check_ref_impl {false};
 
-execution_mode_t execution_mode {execution_mode_t::direct};
-
 int main(int argc, char **argv) {
     using namespace parser;
