+++ tests/benchdnn/utils/parser.cpp
@@ -20,6 +20,7 @@
 #include <unordered_map>
 
 #include "utils/cold_cache.hpp"
+#include "utils/execution_mode.hpp"
 #include "utils/fill.hpp"
 #include "utils/parser.hpp"
 #include "utils/stream_kind.hpp"
